+++ CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 ### Fixed
 - [CSE] Added check for "Resource Type" request parameter. It is only allowed for *CREATE* requests.
 - [CSE] Added missing *creator* attribute in deletion notifications for &lt;crossResourceSubscription> resources.
+- [CSE] Fixed wrong handling of announced &lt;flexContainer> resources. Custom attributes were not announced correctly. Thanks to Alper Ramadan for reporting this issue.